Allow references to hidden file paths
Doorstop excludes hidden files from paths included in its
representation of the version control system files, but this
means that e.g. CI config files in a hidden folder cannot be
included in the references of an item.

This patch allows users to specify file patterns that should not
be hidden in a file (default name: .doorstop_unhidden)