feat(doorstop): Implement external parent
We added an implementation that pull a remote parent
for a doorstop document if not present.

It also validates if you try to import a prefix twice with
different tags and also does recursive download if required.

You need to change your `.doorstop.yml` to add the following

```
external_parent: <git repo url (with .git)>
external_parent_tag: <tag>
settings:
 parent: <PARENT PREFIX>
````